% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{add_year_column}
\alias{add_year_column}
\title{Add year column to data frame if not present}
\usage{
add_year_column(data, table_name)
}
\arguments{
\item{data}{Data frame}

\item{table_name}{Name of the table}
}
\value{
Data frame with YEAR column added
}
\description{
Add year column to data frame if not present
}
