% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{backup_database}
\alias{backup_database}
\title{Backup current database}
\usage{
backup_database(backup_dir = NULL)
}
\arguments{
\item{backup_dir}{Directory to store backup, defaults to user data dir}
}
\value{
Path to backup file
}
\description{
Backup current database
}
