% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_basic_schema}
\alias{check_basic_schema}
\title{Check basic schema requirements}
\usage{
check_basic_schema(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check basic schema requirements
}
