% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_cross_year_data}
\alias{check_cross_year_data}
\title{Check cross-year data consistency}
\usage{
check_cross_year_data(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check cross-year data consistency
}
