% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_data_types}
\alias{check_data_types}
\title{Check data type consistency}
\usage{
check_data_types(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check data type consistency
}
