% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{check_database_age}
\alias{check_database_age}
\title{Check if database needs updating}
\usage{
check_database_age(max_age_days = 90)
}
\arguments{
\item{max_age_days}{Maximum age of database in days before suggesting update}
}
\value{
Logical indicating if update is recommended
}
\description{
Checks if the database should be updated based on age
}
