% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{check_ipeds_updates}
\alias{check_ipeds_updates}
\title{IPEDS Data Update Management}
\usage{
check_ipeds_updates(years = NULL, verbose = TRUE)
}
\arguments{
\item{years}{Vector of years to check, defaults to current year}

\item{verbose}{Whether to print detailed information}
}
\value{
A data frame with information about available data files
}
\description{
Functions for detecting, downloading, and integrating new IPEDS data releases
Check for new IPEDS data releases
}
