% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_null_values}
\alias{check_null_values}
\title{Analyze null values in critical columns}
\usage{
check_null_values(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Analyze null values in critical columns
}
