% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_referential_integrity}
\alias{check_referential_integrity}
\title{Check referential integrity (UNITID consistency)}
\usage{
check_referential_integrity(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check referential integrity (UNITID consistency)
}
