% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_table_exists}
\alias{check_table_exists}
\title{Check if table exists}
\usage{
check_table_exists(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check if table exists
}
