% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_text_encoding}
\alias{check_text_encoding}
\title{Check text encoding issues}
\usage{
check_text_encoding(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check text encoding issues
}
