% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_unitid_validation}
\alias{check_unitid_validation}
\title{Validate UNITID column if present}
\usage{
check_unitid_validation(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Validate UNITID column if present
}
