% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_yaml_available}
\alias{check_yaml_available}
\title{Check if package dependencies are available}
\usage{
check_yaml_available()
}
\description{
Check if package dependencies are available
}
\keyword{internal}
