% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{check_year_consistency}
\alias{check_year_consistency}
\title{Check year consistency in table name vs data}
\usage{
check_year_consistency(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Check result data frame
}
\description{
Check year consistency in table name vs data
}
