% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{clean_ipeds_data}
\alias{clean_ipeds_data}
\title{Clean and standardize IPEDS data}
\usage{
clean_ipeds_data(data, table_name)
}
\arguments{
\item{data}{Data frame to clean}

\item{table_name}{Table name for context}
}
\value{
Cleaned data frame
}
\description{
Clean and standardize IPEDS data
}
