% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{compare_data_versions}
\alias{compare_data_versions}
\title{Compare available updates with current database}
\usage{
compare_data_versions(available_data = NULL, current_data = NULL)
}
\arguments{
\item{available_data}{Data frame from check_ipeds_updates()}

\item{current_data}{Data frame from get_current_database_version()}
}
\value{
A data frame showing what updates are needed
}
\description{
Compare available updates with current database
}
