% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{count_file_rows}
\alias{count_file_rows}
\title{Count rows in a file efficiently}
\usage{
count_file_rows(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
Number of rows
}
\description{
Count rows in a file efficiently
}
