% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{create_empty_dictionary_dataframe}
\alias{create_empty_dictionary_dataframe}
\title{Create empty data frame for dictionary data}
\usage{
create_empty_dictionary_dataframe()
}
\value{
Empty data frame with dictionary structure
}
\description{
Create empty data frame for dictionary data
}
