% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{create_empty_files_dataframe}
\alias{create_empty_files_dataframe}
\title{Create empty data frame with correct structure for files data}
\usage{
create_empty_files_dataframe()
}
\value{
Empty data frame with proper columns
}
\description{
Create empty data frame with correct structure for files data
}
