% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{create_tables_for_year_new}
\alias{create_tables_for_year_new}
\title{Create Tables table for a specific year}
\usage{
create_tables_for_year_new(con, year, table_name, verbose)
}
\description{
Create Tables table for a specific year
}
