% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{cum_inflation}
\alias{cum_inflation}
\title{Cumulative inflation rates}
\usage{
cum_inflation(idbc, start_year, end_year)
}
\arguments{
\item{idbc}{database connector to the IPEDS database}

\item{start_year}{the first year to include}

\item{end_year}{the last year to include}
}
\value{
a list with three elements: annual data, cumulative rate, and average annual rate.
}
\description{
Calculates a cumulative inflation rate for a given time period.
}
\details{
The result is the cumulative inflation rate from the start year to the end year.
The calculation multiplies by one plus the  rate for each year. Finally, one
is subtracted to give a cumlative rate.
}
