% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versioning.R
\name{detect_schema_changes}
\alias{detect_schema_changes}
\title{Compare table schemas between versions}
\usage{
detect_schema_changes(table_name, db_connection = NULL)
}
\arguments{
\item{table_name}{Name of the table to compare}

\item{db_connection}{Database connection}
}
\value{
Data frame with schema differences
}
\description{
Compare table schemas between versions
}
