% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\docType{data}
\name{.ipeds_env}
\alias{.ipeds_env}
\title{Database Management for IPEDSR}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.ipeds_env
}
\description{
Functions to manage the IPEDS database including setup, updates, and connection management.
}
\keyword{datasets}
