% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{download_data_dictionary}
\alias{download_data_dictionary}
\title{Download and parse data dictionary for a specific file}
\usage{
download_data_dictionary(dictionary_link, table_name, verbose = FALSE)
}
\arguments{
\item{dictionary_link}{URL to the dictionary}

\item{table_name}{Name of the table}

\item{verbose}{Whether to print progress}
}
\value{
Data frame with variable definitions
}
\description{
Download and parse data dictionary for a specific file
}
