% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{download_ipeds_file}
\alias{download_ipeds_file}
\title{Download and process a single IPEDS data file}
\usage{
download_ipeds_file(file_info, temp_dir = tempdir(), verbose = TRUE)
}
\arguments{
\item{file_info}{Single row from available data with file information}

\item{temp_dir}{Directory to store temporary files}

\item{verbose}{Whether to print progress}
}
\value{
Path to processed file or NULL if failed
}
\description{
Download and process a single IPEDS data file
}
