% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{extract_data_files_comprehensive}
\alias{extract_data_files_comprehensive}
\title{Extract data file information from IPEDS HTML content using regex patterns}
\usage{
extract_data_files_comprehensive(html_content, year_2digit)
}
\description{
Extract data file information from IPEDS HTML content using regex patterns
}
