% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{extract_dict_files_comprehensive}
\alias{extract_dict_files_comprehensive}
\title{Extract dictionary file information from IPEDS HTML content}
\usage{
extract_dict_files_comprehensive(html_content, year_2digit)
}
\description{
Extract dictionary file information from IPEDS HTML content
}
