% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{extract_table_name_from_link}
\alias{extract_table_name_from_link}
\title{Extract table name from download link}
\usage{
extract_table_name_from_link(link)
}
\arguments{
\item{link}{URL to CSV file}
}
\value{
Table name or empty string
}
\description{
Extract table name from download link
}
