% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{extract_year_from_table_name}
\alias{extract_year_from_table_name}
\title{Extract year from IPEDS table name}
\usage{
extract_year_from_table_name(table_name)
}
\arguments{
\item{table_name}{Name of the table}
}
\value{
Integer year or NA if no year found
}
\description{
Extract year from IPEDS table name
}
