% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_utilities.R
\name{find_unitids}
\alias{find_unitids}
\title{Find UNITID from name}
\usage{
find_unitids(idbc, institution_names, states)
}
\arguments{
\item{idbc}{database connector}

\item{institution_names}{a character vector of distinct institutional names.}

\item{states}{An optional character vector of states, which must be the same length
as the names. States will help names be more specific. Blanks will be ignored.}
}
\value{
A dataframe with the names and UNITIDs where matches were found. Incomplete cases
are dropped.
}
\description{
Use a school's name to attempt a unique ID match
}
\details{
Uses the most recent IPEDS files.
}
