% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{generate_validation_recommendations}
\alias{generate_validation_recommendations}
\title{Generate validation recommendations}
\usage{
generate_validation_recommendations(validation_results)
}
\arguments{
\item{validation_results}{Results from validation}
}
\value{
Vector of recommendations
}
\description{
Generate validation recommendations
}
