% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versioning.R
\name{generate_version_string}
\alias{generate_version_string}
\title{Generate version string based on current date and time}
\usage{
generate_version_string()
}
\value{
Version string
}
\description{
Generate version string based on current date and time
}
