% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{get_admit_funnel}
\alias{get_admit_funnel}
\title{Admit Funnel}
\usage{
get_admit_funnel(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{optional list of UNITIDs to filter to. If NULL, uses configured default_unitid.}

\item{idbc}{database connector}
}
\description{
Admit Funnel
}
