% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletics.R
\name{get_athletics}
\alias{get_athletics}
\title{Get athletics summary data from EADA survey}
\usage{
get_athletics(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{database connector}

\item{UNITIDs}{optional array of UNITIDs to filter to}
}
\value{
A dataframe with Year, UNITID and selected summary statistics from the EADA
data set.
}
\description{
Get athletics summary data from EADA survey
}
