% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athletics.R
\name{get_athletics_sports}
\alias{get_athletics_sports}
\title{Get athletics detailed sports data from EADA survey}
\usage{
get_athletics_sports(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{database connector}

\item{UNITIDs}{optional array of UNITIDs to filter to}
}
\value{
A dataframe with Year, UNITID and selected sports summary statistics from the EADA in long format
data set.
}
\description{
Get athletics detailed sports data from EADA survey
}
