% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_programs.R
\name{get_cip2_counts}
\alias{get_cip2_counts}
\title{CIP distributions}
\usage{
get_cip2_counts(idbc, awlevel = "05", UNITIDs = NULL, first_only = FALSE)
}
\arguments{
\item{idbc}{Database connector for the IPEDS database}

\item{UNITIDs}{optional array of UNITIDs to return, or NULL for everything}

\item{first_only}{When TRUE only counts first majors. Defaults to FALSE.}

\item{level}{The code for the degree level}
}
\value{
A dataframe with Year, UNITID, CIP2, and count, including both first and second majors
}
\description{
Get the distribution of degrees granted by year and institution and CIP2 for
the given degree level.
}
\details{
Codes for the leval are
  3	Associate's degree
  5	Bachelor's degree
  7	Master's degree
  17	Doctor's degree - research/scholarship
  18	Doctor's degree - professional practice
  19	Doctor's degree - other
  20	Certificates of less than 12 weeks
  21	Certificates of at least 12 weeks but less than 1 year
  2	Certificates of at least 1 but less than 2 years
  4	Certificates of at least 2 but less than 4 years
  6	Postbaccalaureate certificate
  8	Post-master's certificate
}
