% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_programs.R
\name{get_cipcodes}
\alias{get_cipcodes}
\title{Get CIP Codes}
\usage{
get_cipcodes(digits = NULL)
}
\arguments{
\item{digits}{optional number of digits to truncate to, or NULL for no truncation}

\item{idbc}{Database connector}
}
\value{
A data frame with CIPCODE, Subject
}
\description{
Get the latest CIP code table from IPEDS
}
\details{
If CIPs are truncated, there are multiple versions, which are reduced
through distinct(). That may not be optimal.
}
