% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_programs.R
\name{get_cips}
\alias{get_cips}
\title{Get CIPS}
\usage{
get_cips(UNITIDs = NULL, years = NULL, cip_codes = NULL, awlevel = "05")
}
\arguments{
\item{UNITIDs}{UNITIDs, or NULL for everything}

\item{years}{The years of data to retrieve; the table name minus one. Years before 2007 are invalid.}

\item{cip_codes}{Provide a vector of codes to search for. This can be two-digit, four, or six. Note
code "99" is all degrees, which is filtered out if you leave cip_codes NULL.}

\item{awlevel}{Award level, defaults to "05" for Bachelors}

\item{idbc}{Database connector}
}
\value{
A data frame with UNITID, CIPCODE, MAJORNUM, N (the count of completions), and Year
}
\description{
Given institutional IDs, get completions by year and CIP for each
}
