% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{get_cohort_stats}
\alias{get_cohort_stats}
\title{Get cohort retention and graduation history}
\usage{
get_cohort_stats(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{an array of IDs or NULL for everything (default)}
}
\value{
Cohort year and size, and rates and numbers of returning and
graduating students.
}
\description{
Get cohort retention and graduation history
}
\details{
This report omits Year because it's keyed on Chort.
The columns Yr1 through Yr6 estimate enrollment.
}
