% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{get_comprehensive_file_listing}
\alias{get_comprehensive_file_listing}
\title{Get comprehensive file listing for multiple years}
\usage{
get_comprehensive_file_listing(
  years = 2020:2024,
  include_dictionaries = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{years}{Vector of years to process}

\item{include_dictionaries}{Whether to download dictionary information}

\item{verbose}{Whether to print progress}
}
\value{
Data frame with all file information
}
\description{
Get comprehensive file listing for multiple years
}
