% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config_path}
\alias{get_config_path}
\title{IPEDSR Configuration Management}
\usage{
get_config_path()
}
\value{
Character string with the config file path
}
\description{
Functions for managing user-specific configuration settings
Get the path to the configuration file
}
\keyword{internal}
