% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{get_current_database_version}
\alias{get_current_database_version}
\title{Get current database version information}
\usage{
get_current_database_version()
}
\value{
A data frame with current database metadata
}
\description{
Get current database version information
}
