% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_personnel.R
\name{get_employees}
\alias{get_employees}
\title{Get employees}
\usage{
get_employees(idbc, UNITIDs = 218070)
}
\arguments{
\item{idbc}{A database connector}

\item{UNITIDs}{an array of UNITIDs, defaulting to Furman}
}
\value{
a dataframe with employee information
}
\description{
Get employees
}
