% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_expected_ipeds_type}
\alias{get_expected_ipeds_type}
\title{Get expected data type for known IPEDS columns}
\usage{
get_expected_ipeds_type(column_name)
}
\arguments{
\item{column_name}{Name of the column}
}
\value{
Expected type or NULL
}
\description{
Get expected data type for known IPEDS columns
}
