% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_expected_range}
\alias{get_expected_range}
\title{Get expected range for known IPEDS columns}
\usage{
get_expected_range(column_name)
}
\arguments{
\item{column_name}{Name of the column}
}
\value{
List with min and max or NULL
}
\description{
Get expected range for known IPEDS columns
}
