% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{get_fa_info}
\alias{get_fa_info}
\title{Get financial aid}
\usage{
get_fa_info(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{IPEDS database connection}

\item{UNITIDs}{Optional list of UNITIDs to filter to}
}
\description{
Get financial aid
}
