% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_personnel.R
\name{get_faculty}
\alias{get_faculty}
\title{Get IPEDS faculty counts}
\usage{
get_faculty(UNITIDs = NULL, before_2011 = FALSE)
}
\arguments{
\item{UNITIDs}{IPEDS school IDs. If NULL, uses configured default_unitid.}

\item{before_2011}{If true, includes old data}
}
\value{
A dataframe with UNITID, Year, Tuition, Fees, RoomBoard for undergrad
}
\description{
Get IPEDS faculty counts
}
\details{
The faculty status codes for rank and tenure were different before 2011, making
comparisons more difficult. By default, those records are omitted.
}
