% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_financials.R
\name{get_finances}
\alias{get_finances}
\title{Get financials}
\usage{
get_finances(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{vector of UNITIDs to retrieve. If NULL, uses configured default_unitid.}
}
\value{
Dataframe with endowment, revenue and other statistics, ordered by UNITID and Year.
}
\description{
Get financials
}
