% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_financials.R
\name{get_finances_public}
\alias{get_finances_public}
\title{Get financials for public institutions}
\usage{
get_finances_public(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{Database connector}

\item{UNTIDs}{vector of UNITIDs to retrieve}
}
\value{
Dataframe with endowment, revenue and other statistics
}
\description{
Get financials for public institutions
}
