% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_completions.R
\name{get_grad_pell_rates}
\alias{get_grad_pell_rates}
\title{Get pell grad rates}
\usage{
get_grad_pell_rates(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{Database connector}

\item{UNITIDs}{Array of identifiers}
}
\value{
Dataframe with year, unitid, and four year rates having column names
that identify the demographic (total means everyone). Only returns data after
2015. Rates are as of August 31 of the identified year.
}
\description{
Get six-year graduation rates for cohort undergraduates for
Pell and Stafford students.
}
