% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_completions.R
\name{get_grad_rates}
\alias{get_grad_rates}
\title{Get grad rates}
\usage{
get_grad_rates(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{Array of identifiers, or NULL for all schools}

\item{idbc}{Database connector}
}
\value{
Dataframe with year and cohort and cohort size, unitid, 6 year cumulative grad rate,
and 4-, 5-, and 6- year rates.
}
\description{
Get grad rates
}
\details{
The four, five, and six year rates are all for the same cohort, and
all reported at the same time, which is six years after entry. The four year rate
means four or less years, but five and six mean exactly those many years.
}
