% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_characteristics.R
\name{get_ids}
\alias{get_ids}
\title{Get IPEDS UNITIDs for institutions matching criteria}
\usage{
get_ids(idbc, year = NULL, control = NULL, states = NULL, level = NULL)
}
\arguments{
\item{idbc}{A database connector}

\item{year}{The year to get, or NULL (default) for the most recent}

\item{control}{Optional array of CONTROL codes to filter to}

\item{states}{Optional array of state abbreviations to filter to}

\item{level}{Optional array of ICLEVEL codes to filter to}
}
\value{
A vector with UNITIDs matching the criteria
}
\description{
Get IPEDS UNITIDs for institutions matching criteria
}
\details{
State abbreviations are the standard postal codes.
CONTROL codes are:
 1	Public
 2	Private not-for-profit
 3	Private for-profit
 -3	{Not available}
 ICLEVEL codes are:
 1	Four or more years
 2	At least 2 but less than 4 years
 3	Less than 2 years (below associate)
 -3	{Not available}
 You can provide an array like c(1,3) to get multiple levels or controls.
 If left blank, all levels/controls/states are included.
}
