% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{get_integration_test_suite}
\alias{get_integration_test_suite}
\title{Get the integration test suite}
\usage{
get_integration_test_suite(quick_mode = FALSE)
}
\arguments{
\item{quick_mode}{Whether to use abbreviated test suite}
}
\value{
Named list of test functions
}
\description{
Get the integration test suite
}
