% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_completions.R
\name{get_ipeds_completions}
\alias{get_ipeds_completions}
\title{Get completions by school and CIP}
\usage{
get_ipeds_completions(years, UNITIDs = NULL, awlevel = "05")
}
\arguments{
\item{years}{The years of data to retrieve; the table name minus one. Years before 2007 are invalid.}

\item{UNITIDs}{if NULL, uses configured default_unitid}

\item{awlevel}{Award level, defaults to "05" for Bachelors}
}
\value{
a dataframe with institutional characteristics, year, and graduates by CIP
}
\description{
Get completions by school and CIP
}
\details{
The CIP code is 2-digit for economy of results
}
