% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{get_ipeds_connection}
\alias{get_ipeds_connection}
\title{Get database connection}
\usage{
get_ipeds_connection(read_only = TRUE)
}
\arguments{
\item{read_only}{Logical, if TRUE opens connection in read-only mode}
}
\value{
DBI connection object
}
\description{
Internal function to get a database connection
}
\keyword{internal}
