% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{get_ipeds_downloads_path}
\alias{get_ipeds_downloads_path}
\title{Get persistent downloads directory path}
\usage{
get_ipeds_downloads_path()
}
\value{
Path to downloads directory
}
\description{
Gets the path to the persistent downloads directory (same location as database)
}
