% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_personnel.R
\name{get_ipeds_faculty_salaries}
\alias{get_ipeds_faculty_salaries}
\title{Faculty salaries}
\usage{
get_ipeds_faculty_salaries(UNITIDs = NULL, years = NULL)
}
\arguments{
\item{UNITIDs}{an array of UNITIDs, or NULL. If NULL, uses configured default_unitid.}

\item{years}{an array of years to pull}
}
\value{
a dataframe with faculty salaries by year and UNITID
}
\description{
Faculty salaries
}
