% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_utilities.R
\name{get_ipeds_table}
\alias{get_ipeds_table}
\title{Retrieve IPEDS data with columns renamed}
\usage{
get_ipeds_table(idbc, table_name, year2, UNITIDs = NULL)
}
\arguments{
\item{idbc}{Database connection}

\item{table_name}{The complete name of the table including year}

\item{year2}{A two-character year, as in 20XX, so 2020 is "20" and 2021 is "21"}

\item{UNITIDs}{an optional list of IDs, or NULL for everything}
}
\value{
A dataframe with human-readable data.
}
\description{
get a whole table with the columns named for their descriptions,
and any code values replaced with the text version.
}
\details{
The year2 input is used to find the var and value tables needed
         to upack the codes into text
}
