% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_ipedsr_config}
\alias{get_ipedsr_config}
\title{Get IPEDSR configuration}
\usage{
get_ipedsr_config(key = NULL, default = NULL)
}
\arguments{
\item{key}{Optional specific configuration key to retrieve. If NULL, returns all configuration.}

\item{default}{Default value to return if key is not found}
}
\value{
Configuration value(s)
}
\description{
Get IPEDSR configuration
}
\examples{
\dontrun{
# Get all configuration
get_ipedsr_config()

# Get specific value
get_ipedsr_config("default_unitid")

# Get with default if not set
get_ipedsr_config("default_unitid", default = 218070)
}
}
