% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{get_retention}
\alias{get_retention}
\title{Get cohort size and retention rates}
\usage{
get_retention(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{database connector}

\item{UNITIDs}{optional IDs to retrieve, or NULL for everything}
}
\value{
a dataframe with UNITID, Year, Retention, CohortSize
}
\description{
Get cohort size and retention rates
}
\details{
The year has been adjusted so that the cohort size and retention
rate are for the same group of students with cohort = year. A cohort column
is included to make this clear.
}
