% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versioning.R
\name{get_table_schema}
\alias{get_table_schema}
\title{Get current table schema}
\usage{
get_table_schema(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
Data frame with schema information
}
\description{
Get current table schema
}
