% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_versioning.R
\name{get_table_statistics}
\alias{get_table_statistics}
\title{Get statistics for a database table}
\usage{
get_table_statistics(table_name, db_connection)
}
\arguments{
\item{table_name}{Name of the table}

\item{db_connection}{Database connection}
}
\value{
List with table statistics
}
\description{
Get statistics for a database table
}
