% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_financials.R
\name{get_tuition}
\alias{get_tuition}
\title{Get IPEDS tuition}
\usage{
get_tuition(idbc, UNITIDs = NULL)
}
\arguments{
\item{idbc}{Database connector}

\item{UNITIDs}{IPEDS school IDs. If NULL, gets everything}
}
\value{
A dataframe with UNITID, Year, Tuition, Fees, RoomBoard for undergrad
}
\description{
Get IPEDS tuition
}
