% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{get_validation_checks}
\alias{get_validation_checks}
\title{Get list of validation checks for a given level}
\usage{
get_validation_checks(validation_level)
}
\arguments{
\item{validation_level}{Level of validation}
}
\value{
Vector of check names
}
\description{
Get list of validation checks for a given level
}
