% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_utilities.R
\name{get_valueset}
\alias{get_valueset}
\title{Get valueset}
\usage{
get_valueset(idbc, my_table)
}
\arguments{
\item{idbc}{database connector to IPEDS db}

\item{my_table}{The IPEDS table being accessed}
}
\description{
Given a table name
look up the variables, codes, and labels that go with codes
(e.g. STABBR has a code "SC" for "South Carolina")
}
