% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{import_data_file_new}
\alias{import_data_file_new}
\title{Import a data file to database}
\usage{
import_data_file_new(file_path, table_name, con, verbose)
}
\description{
Import a data file to database
}
