% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{import_year_dictionaries_new}
\alias{import_year_dictionaries_new}
\title{Import dictionary files for a specific year}
\usage{
import_year_dictionaries_new(downloads_dir, con, year, verbose)
}
\description{
Import dictionary files for a specific year
}
