% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\docType{data}
\name{ipeds_available_files}
\alias{ipeds_available_files}
\title{Available shorthands for IPEDS data files}
\format{
An object of class \code{character} of length 6.
}
\usage{
ipeds_available_files
}
\description{
Available shorthands for IPEDS data files
}
\keyword{datasets}
