% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ipeds_download}
\alias{ipeds_download}
\title{Download IPEDS data files from NCES}
\usage{
ipeds_download(
  data_files,
  data_dir = "ipeds",
  force = FALSE,
  sleep_interval = 1
)
}
\description{
Download IPEDS data files from NCES
}
