% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ipeds_files}
\alias{ipeds_files}
\title{Get filenames corresponding to an IPEDS survey component}
\usage{
ipeds_files(survey, years)
}
\description{
Get filenames corresponding to an IPEDS survey component
}
