% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{ipeds_get_enrollment}
\alias{ipeds_get_enrollment}
\title{Get IPEDS enrollment}
\usage{
ipeds_get_enrollment(UNITIDs = NULL, StudentTypeCode = 1)
}
\arguments{
\item{UNITIDs}{The schools to retrieve, or NULL (default) for all.}

\item{StudentTypeCode}{One or more numeric codes to specify what type of student
to filter to. See Details for, well, details. Defaults to 1, all students.}

\item{idbc}{Database connector for the IPEDS database.}
}
\value{
A dataframe with UNITID, Year, Total, Men, Women, White, Black, Hispanic, and NRAlien.
}
\description{
Use the efYEAR tables to retrieve student counts.
}
\details{
The StudentTypeCode is one or more of
1	= All students total
2	= All students, Undergraduate total
3	= All students, Undergraduate, Degree/certificate-seeking total
4	= All students, Undergraduate, Degree/certificate-seeking, First-time
5	= All students, Undergraduate, Other degree/certificate-seeking
19	= All students, Undergraduate, Other degree/certifcate-seeking, Transfer-ins
20	= All students, Undergraduate, Other degree/certifcate-seeking, Continuing
11	= All students, Undergraduate, Non-degree/certificate-seeking
12	= All students, Graduate
21	= Full-time students total
22	= Full-time students, Undergraduate total
23	= Full-time students, Undergraduate, Degree/certificate-seeking total
24	= Full-time students, Undergraduate, Degree/certificate-seeking, First-time
25	= Full-time students, Undergraduate, Degree/certificate-seeking, Other degree/certificate-seeking
39	= Full-time students, Undergraduate, Other degree/certifcate-seeking, Transfer-ins
40	= Full-time students, Undergraduate, Other degree/certifcate-seeking, Continuing
31	= Full-time students, Undergraduate, Non-degree/certificate-seeking
32	= Full-time students, Graduate
41	= Part-time students total
42	= Part-time students, Undergraduate total
43	= Part-time students, Undergraduate, Degree/certificate-seeking total
44	= Part-time students, Undergraduate, Degree/certificate-seeking, First-time
45	= Part-time students, Undergraduate, Degree/certificate-seeking, Other degree/certificate-seeking
59	= Part-time students, Undergraduate, Other degree/certifcate-seeking, Transfer-ins
60	= Part-time students, Undergraduate, Other degree/certifcate-seeking, Continuing
51	= Part-time students, Undergraduate, Non-degree/certificate-seeking
52	= Part-time students, Graduate
}
