% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{ipeds_load_series}
\alias{ipeds_load_series}
\title{Load multiple years of an IPEDS survey component}
\usage{
ipeds_load_series(
  data_files,
  data_dir = "ipeds",
  force = FALSE,
  read = NULL,
  ...
)
}
\description{
Load multiple years of an IPEDS survey component
}
