% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_utilities.R
\name{my_dbListTables}
\alias{my_dbListTables}
\title{Find matching tables}
\usage{
my_dbListTables(search_string)
}
\arguments{
\item{search_string}{A regex pattern to search for in the table names (case-sensitive)}
}
\value{
A character vector of table names that match the search string
}
\description{
Find matching tables
}
