% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_scraping.R
\name{parse_dictionary_content}
\alias{parse_dictionary_content}
\title{Parse data dictionary content}
\usage{
parse_dictionary_content(content, table_name)
}
\arguments{
\item{content}{Raw content from dictionary file}

\item{table_name}{Name of the table}
}
\value{
Data frame with variable definitions
}
\description{
Parse data dictionary content
}
