% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{print_config}
\alias{print_config}
\title{Print configuration in a readable format}
\usage{
print_config(config)
}
\arguments{
\item{config}{Configuration list}
}
\description{
Print configuration in a readable format
}
\keyword{internal}
