% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{print_integration_test_summary}
\alias{print_integration_test_summary}
\title{Print integration test summary}
\usage{
print_integration_test_summary(test_results)
}
\arguments{
\item{test_results}{Results from integration tests}
}
\description{
Print integration test summary
}
