% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{process_data_files_new}
\alias{process_data_files_new}
\title{Process (download and import) data files}
\usage{
process_data_files_new(data_files, downloads_dir, con, force_download, verbose)
}
\description{
Process (download and import) data files
}
