% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{process_dict_files_new}
\alias{process_dict_files_new}
\title{Process (download and import) dictionary files}
\usage{
process_dict_files_new(
  dict_files,
  downloads_dir,
  con,
  year,
  force_download,
  verbose
)
}
\description{
Process (download and import) dictionary files
}
