% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{process_year_data}
\alias{process_year_data}
\title{Process data and dictionary files for a single year}
\usage{
process_year_data(year, downloads_dir, con, force_download, verbose)
}
\description{
Process data and dictionary files for a single year
}
