% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{restore_database_backup}
\alias{restore_database_backup}
\title{Restore database from backup}
\usage{
restore_database_backup(backup_path, confirm = TRUE)
}
\arguments{
\item{backup_path}{Path to backup file to restore}

\item{confirm}{Whether to require confirmation before restoring}
}
\description{
Restore database from backup
}
