% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{run_single_integration_test}
\alias{run_single_integration_test}
\title{Run a single integration test}
\usage{
run_single_integration_test(test_function, test_year, verbose = TRUE)
}
\arguments{
\item{test_function}{Function to run}

\item{test_year}{Year for testing}

\item{verbose}{Whether to show detailed output}
}
\value{
Test result list
}
\description{
Run a single integration test
}
