% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{scrape_ipeds_releases}
\alias{scrape_ipeds_releases}
\title{Scrape IPEDS data file information from NCES website}
\usage{
scrape_ipeds_releases(year, verbose = FALSE)
}
\arguments{
\item{year}{The year to scrape data for}

\item{verbose}{Whether to print progress messages}
}
\value{
A data frame with file information
}
\description{
Scrape IPEDS data file information from NCES website
}
