% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_ipeds}
\alias{setup_ipeds}
\title{Set up the IPEDS DuckDB database}
\usage{
setup_ipeds(url = "https://yourserver.org/ipeds.duckdb")
}
\arguments{
\item{url}{Where to get the duckdb IPEDS database}
}
\value{
a zero upon success
}
\description{
Downloads the IPEDS database file to a standard location and sets
an environment variable in .Renviron so it can be accessed from any project.
}
