% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{show_detailed_validation_results}
\alias{show_detailed_validation_results}
\title{Show detailed validation results for problematic tables}
\usage{
show_detailed_validation_results(validation_results)
}
\arguments{
\item{validation_results}{Results from validation}
}
\description{
Show detailed validation results for problematic tables
}
