% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{store_database_metadata}
\alias{store_database_metadata}
\title{Store database metadata}
\usage{
store_database_metadata()
}
\description{
Stores information about when the database was downloaded/updated
}
\keyword{internal}
