% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{table_exists_in_db_new}
\alias{table_exists_in_db_new}
\title{Helper functions for the new update system}
\usage{
table_exists_in_db_new(con, table_name)
}
\description{
Helper functions for the new update system
}
