% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_data_integrity_checks}
\alias{test_data_integrity_checks}
\title{Test data integrity checks (comprehensive test only)}
\usage{
test_data_integrity_checks(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test data integrity checks (comprehensive test only)
}
