% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_data_processing_functionality}
\alias{test_data_processing_functionality}
\title{Test data processing functionality}
\usage{
test_data_processing_functionality(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test data processing functionality
}
