% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_database_connection}
\alias{test_database_connection}
\title{Test database connection functionality}
\usage{
test_database_connection(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test database connection functionality
}
