% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_file_download_functionality}
\alias{test_file_download_functionality}
\title{Test file download functionality}
\usage{
test_file_download_functionality(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test file download functionality
}
