% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_user_interface_functions}
\alias{test_user_interface_functions}
\title{Test user interface functions}
\usage{
test_user_interface_functions(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test user interface functions
}
