% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_validation_system}
\alias{test_validation_system}
\title{Test validation system}
\usage{
test_validation_system(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test validation system
}
