% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration_testing.R
\name{test_web_scraping_functionality}
\alias{test_web_scraping_functionality}
\title{Test web scraping functionality}
\usage{
test_web_scraping_functionality(test_year, verbose = FALSE)
}
\arguments{
\item{test_year}{Year for testing}

\item{verbose}{Verbose output}
}
\value{
Test result
}
\description{
Test web scraping functionality
}
