% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{ui_backup_database}
\alias{ui_backup_database}
\title{Backup database}
\usage{
ui_backup_database(interactive = TRUE)
}
\arguments{
\item{interactive}{Whether to use interactive prompts}
}
\description{
Backup database
}
