% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{ui_restore_database}
\alias{ui_restore_database}
\title{Restore database from backup}
\usage{
ui_restore_database(interactive = TRUE)
}
\arguments{
\item{interactive}{Whether to use interactive prompts}
}
\description{
Restore database from backup
}
