% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_interface.R
\name{ui_show_status}
\alias{ui_show_status}
\title{Show database status and summary}
\usage{
ui_show_status(interactive = TRUE)
}
\arguments{
\item{interactive}{Whether to use interactive prompts}
}
\description{
Show database status and summary
}
