% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{update_consolidated_dictionaries_new}
\alias{update_consolidated_dictionaries_new}
\title{Update the consolidated dictionary tables with new year data}
\usage{
update_consolidated_dictionaries_new(con, verbose)
}
\description{
Update the consolidated dictionary tables with new year data
}
