% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_updates.R
\name{update_data}
\alias{update_data}
\title{Update IPEDS data for specified years}
\usage{
update_data(
  years = NULL,
  force_download = FALSE,
  verbose = TRUE,
  backup_first = TRUE
)
}
\arguments{
\item{years}{Vector of years to update (4-digit integers), defaults to current year}

\item{force_download}{Logical, if TRUE will re-download even if files exist}

\item{verbose}{Logical, if TRUE prints detailed progress messages}

\item{backup_first}{Logical, if TRUE creates database backup before updates}
}
\value{
List with summary of what was updated
}
\description{
Update IPEDS data for specified years
}
