% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{update_ipeds_database}
\alias{update_ipeds_database}
\title{Update IPEDS database}
\usage{
update_ipeds_database(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical, if TRUE suppresses progress messages}
}
\value{
Logical indicating success
}
\description{
Forces an update of the IPEDS database
}
