% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{view_ipedsr_config}
\alias{view_ipedsr_config}
\title{View current IPEDSR configuration}
\usage{
view_ipedsr_config()
}
\value{
Invisibly returns the configuration list
}
\description{
View current IPEDSR configuration
}
\examples{
\dontrun{
view_ipedsr_config()
}
}
