% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_characteristics.R
\name{get_characteristics}
\alias{get_characteristics}
\title{Get institutional characteristics}
\usage{
get_characteristics(idbc, year = NULL, UNITIDs = NULL)
}
\arguments{
\item{idbc}{A database connector}

\item{year}{The year to get, or NULL (default) for the most recent year}

\item{UNITIDs}{optional array of UNITIDs to filter to}
}
\value{
a dataframe with institutional characteristics
}
\description{
Get institutional characteristics
}
\details{
Apparently these tables include 'blob' columns that won't allow downloading
of the full thing. See https://github.com/r-dbi/odbc/issues/309.
So we have to select around those.
}
