% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_personnel.R
\name{get_employees}
\alias{get_employees}
\title{Get employees}
\usage{
get_employees(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{an array of UNITIDs, or NULL. If NULL, uses configured default_unitid.}
}
\value{
a dataframe with employee information
}
\description{
Get employees
}
