% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_management.R
\name{get_ipeds_db_path}
\alias{get_ipeds_db_path}
\title{Get IPEDS database path}
\usage{
get_ipeds_db_path()
}
\value{
Character string with the database path
}
\description{
Returns the path where the IPEDS database should be stored
}
\keyword{internal}
