% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_utilities.R
\name{get_labels}
\alias{get_labels}
\title{Convert to friendly names}
\usage{
get_labels(my_table, df, verbose = FALSE)
}
\arguments{
\item{my_table}{The IPEDS table being accessed}

\item{df}{Data from the table given}

\item{verbose}{Logical, defaults to FALSE--Include the long descriptions of vars?}
}
\description{
Given results from an IPEDS table,
create a long dataframe with friendly variable names
and translated code values.
}
