% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_cohorts.R
\name{get_retention}
\alias{get_retention}
\title{Get cohort size and retention rates}
\usage{
get_retention(UNITIDs = NULL)
}
\arguments{
\item{UNITIDs}{optional IDs to retrieve, or NULL for everything}

\item{idbc}{database connector}
}
\value{
a dataframe with UNITID, Year, Retention, CohortSize
}
\description{
Get cohort size and retention rates
}
\details{
The cohort  = year - 1 since the retention rate is reported a year
later. A cohort column is included to make this clear.
}
